+++ lib/app/modules/login/controllers/login_controller.dart
@@ -5,7 +5,9 @@ class LoginController extends GetxController {
   final emailController = TextEditingController();
   final passwordController = TextEditingController();
   final formKey = GlobalKey<FormState>();
-  final passwordInvisible = true.obs;
+  final _isPasswordVisible = false.obs;
+
+  bool get isPasswordVisible => _isPasswordVisible.value;
 
   /// Logs the user in.
   Future<void> login() async {
@@ -21,35 +23,9 @@ class LoginController extends GetxController {
   }
 
   void togglePasswordInvisible() {
-    passwordInvisible.value = !passwordInvisible.value;
+    _isPasswordVisible.value = !_isPasswordVisible.value;
   }
-
-  /// Validates the email.
-  String? validateEmail(String? email) {
-    if (email == null || email.isEmpty) {
-      return 'Email is required.';
-    }
-
-    if (!GetUtils.isEmail(email)) {
-      return 'Please enter a valid email address.';
-    }
-
-    return null;
-  }
-
-  /// Validates the password.
-  String? validatePassword(String? password) {
-    if (password == null || password.isEmpty) {
-      return 'Password is required.';
-    }
-
-    if (password.length < 6) {
-      return 'Password must be at least 6 characters long.';
-    }
-
-    return null;
-  }
-
+  
   /// Simulates a login request. (Delete this after implementing supabase auth.)
   Future<void> _login(String email, String password) async {
     print('Email: $email');
