+++ lib/app/modules/sign_up/controllers/sign_up_controller.dart
@@ -6,7 +6,9 @@ class SignUpController extends GetxController {
   final passwordController = TextEditingController();
   final passwordConfirmController = TextEditingController();
   final formKey = GlobalKey<FormState>();
-  final passwordInvisible = true.obs;
+  final _isPasswordVisible = false.obs;
+
+  bool get isPasswordVisible => _isPasswordVisible.value;
 
   /// Register the user.
   Future<void> signUp() async {
@@ -23,7 +25,7 @@ class SignUpController extends GetxController {
   }
 
   void togglePasswordInvisible() {
-    passwordInvisible.value = !passwordInvisible.value;
+    _isPasswordVisible.value = !_isPasswordVisible.value;
   }
 
   Future<void> _signUp(
@@ -37,40 +39,15 @@ class SignUpController extends GetxController {
 
     await Future<dynamic>.delayed(const Duration(seconds: 1));
   }
+  // String? validatePasswordConfirm(String? passwordConfirm) {
+  //   if (passwordConfirm == null || passwordConfirm.isEmpty) {
+  //     return 'Password confirmation is required.';
+  //   }
 
-  String? validateEmail(String? email) {
-    if (email == null || email.isEmpty) {
-      return 'Email is required.';
-    }
+  //   if (passwordConfirm != passwordController.text) {
+  //     return 'Passwords do not match.';
+  //   }
 
-    if (!GetUtils.isEmail(email)) {
-      return 'Please enter a valid email address.';
-    }
-
-    return null;
-  }
-
-  String? validatePassword(String? password) {
-    if (password == null || password.isEmpty) {
-      return 'Password is required.';
-    }
-
-    if (password.length < 6) {
-      return 'Password must be at least 6 characters.';
-    }
-
-    return null;
-  }
-
-  String? validatePasswordConfirm(String? passwordConfirm) {
-    if (passwordConfirm == null || passwordConfirm.isEmpty) {
-      return 'Password confirmation is required.';
-    }
-
-    if (passwordConfirm != passwordController.text) {
-      return 'Passwords do not match.';
-    }
-
-    return null;
-  }
+  //   return null;
+  // }
 }
